+++ include/uit/mpi/group_utils.hpp
@@ -8,7 +8,7 @@
 #include "../../../third-party/Empirical/source/base/assert.h"
 #include "../../../third-party/Empirical/source/tools/math.h"
 
-#include "../utility/math_utils.hpp"
+#include "../math/math_utils.hpp"
 #include "../utility/print_utils.hpp"
 #include "../debug/safe_compare.hpp"
