+++ include/uit/parallel/thread_utils.hpp
@@ -8,7 +8,7 @@
 #include "../../../third-party/Empirical/source/base/assert.h"
 
 #include "../utility/exec_utils.hpp"
-#include "../utility/math_utils.hpp"
+#include "../math/math_utils.hpp"
 
 namespace uit {
