+++ microbenchmarks/conduit/DuctBenchmarkRegistration.hpp
@@ -25,7 +25,7 @@
 #include "uit/meta/Curry.hpp"
 #include "uit/meta/ForEach.hpp"
 #include "uit/debug/safe_cast.hpp"
-#include "uit/utility/stats_utils.hpp"
+#include "uit/math/stats_utils.hpp"
 #include "uit/utility/TimeGuard.hpp"
 #include "uit/parallel/ThreadTeam.hpp"
 #include "uit/parallel/thread_utils.hpp"
