+++ tests/conduit/intra/IntraDuct.hpp
@@ -17,7 +17,7 @@
 #include "uit/topology/DyadicTopologyFactory.hpp"
 #include "uit/utility/assign_utils.hpp"
 #include "uit/utility/CircularIndex.hpp"
-#include "uit/utility/math_utils.hpp"
+#include "uit/math/math_utils.hpp"
 #include "uit/debug/safe_cast.hpp"
 #include "uit/debug/safe_compare.hpp"
