+++ tests/conduit/proc/ProcDuct.hpp
@@ -24,7 +24,7 @@
 #include "uit/topology/RingTopologyFactory.hpp"
 #include "uit/utility/assign_utils.hpp"
 #include "uit/utility/CircularIndex.hpp"
-#include "uit/utility/math_utils.hpp"
+#include "uit/math/math_utils.hpp"
 #include "uit/debug/safe_cast.hpp"
 
 const uit::MpiGuard guard;
